% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyzer.R
\name{Analyzer}
\alias{Analyzer}
\title{Analyzer of an artificial Neural Network}
\description{
This class analyzes a passed neural network and stores its internal structures
and layers independently of the actual class of the network. With the help
of this object, various methods of interpretable machine learning can be applied
to it for a better understanding of individual predictions or of the whole model.
You can use models from the following libraries:
\itemize{
\item \code{\link[keras]{keras}},
\item \code{\link[neuralnet]{neuralnet}}
}
\subsection{Implemented methods}{

Methods that aim to explain an individual prediction of a neural network are
called \strong{local}. In contrast, \strong{global} methods provide an explanation for the entire
model. An object of the Analyzer class can be applied to the following local
and global methods:
\itemize{
\item Global:
\itemize{
\item \link{Connection_Weights}, Olden et al. (2004)
}
\item Local:
\itemize{
\item Layerwise Relevance Propagation (\link{LRP}), Bach et al. (2015)
\item Deep Learning Important Feartures (\link{DeepLift}), Shrikumar et al. (2017)
\item Vanilla \link{Gradient}
\item \link{SmoothGrad}
}
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The given neural network.}

\item{\code{input_dim}}{Dimension of the input features.}

\item{\code{input_names}}{Names of the input features}

\item{\code{output_dim}}{Dimension of the models output, i.e. dimension of the
response variables.}

\item{\code{output_names}}{A list of names for the response variables.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Analyzer$new()}}
\item \href{#method-clone}{\code{Analyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new analyzer for a given neural network.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyzer$new(
  model,
  feature_names = NULL,
  response_names = NULL,
  dtype = "float"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A trained neural network for classification or regression
tasks to be interpreted. Only models from the following types or packages
are allowed: \code{\link[keras]{keras_model}},
\code{\link[keras]{keras_model_sequential}} or
\code{\link[neuralnet]{neuralnet}}.}

\item{\code{feature_names}}{A list of names for the input features. Use the
default value \code{NULL} for the default names (X1, X2, ...).}

\item{\code{response_names}}{A list of names for the response variables. Use the
default value \code{NULL} for the default names (Y1, Y2, ...).}

\item{\code{dtype}}{The data type for the calculations. Use either \code{'float'} or \code{'double'}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new instance of the R6 class \code{'Analyzer'}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Analyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
