% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection_Weights.R
\name{func_connection_weights}
\alias{func_connection_weights}
\title{Connection Weights method}
\usage{
func_connection_weights(layers, out_class = NULL)
}
\arguments{
\item{layers}{List of layers of type \code{\link{Dense_Layer}}.}

\item{out_class}{If the given model is a classification model, this
parameter can be used to determine which class the importance score should be
calculated for. Use the default value \code{NULL} to return the importance
for all classes.}
}
\value{
If \code{out_class} is \code{NULL} it returns a matrix of shape \emph{(in, out)},
which contains the importance scores for each input variable to the
output predictions. Otherwise returns a vector of the importance scores
for each input variable for the given output class.
}
\description{
This function implements the \emph{Connection Weight} method investigated by
Olden et al. (2004) which results in a feature importance score for each input
variable. The basic idea is to multiply up all path weights for each
possible connection between an input feature and the output and then
calculate the sum over them. For a neural network with \eqn{3} hidden layers with weight
matrices \eqn{W_1}, \eqn{W_2} and \eqn{W_3} this method results in a simple
matrix multiplication
\deqn{W_1 \cdot W_2 \cdot W_3. }
}
\examples{
# create dense layers
W_1 <- matrix(rnorm(3*10), nrow = 3, ncol = 10)
b_1 <- rnorm(10)
W_2 <- matrix(rnorm(10*5), nrow = 10, ncol = 5)
b_2 <- rnorm(5)
W_3 <- matrix(rnorm(5*2), nrow = 5, ncol = 2)
b_3 <- rnorm(2)

dense_1 <- Dense_Layer$new(W_1, b_1, get_activation("relu"))
dense_2 <- Dense_Layer$new(W_2, b_2, get_activation("relu"))
dense_3 <- Dense_Layer$new(W_3, b_3, get_activation("softmax"))

# calculate importance scores for class 1
func_connection_weights(list(dense_1, dense_2, dense_3), out_class = 1)

# calculate importance scores for all classes
func_connection_weights(list(dense_1, dense_2, dense_3), out_class = NULL)

}
\references{
J. D. Olden et al. (2004) \emph{An accurate comparison of methods for
quantifying variable importance in artificial neural networks using
simulated data.} Ecological Modelling 178, p. 389–397
}
\seealso{
\code{\link{Analyzer}}, \code{\link{func_deeplift}}, \code{\link{func_lrp}}
}
