% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradient.R
\name{func_gradient}
\alias{func_gradient}
\title{Gradient method}
\usage{
func_gradient(layers, out_class = NULL)
}
\arguments{
\item{layers}{List of layers of type \code{\link{Dense_Layer}}.}

\item{out_class}{If the given model is a classification model, this
parameter can be used to determine which class the gradients should be
calculated for. Use the default value \code{NULL} to return the gradients
for all classes.}
}
\value{
If \code{out_class} is \code{NULL} it returns a matrix of shape \emph{(in, out)},
which contains the gradients for each input variable to the
output predictions. Otherwise returns a vector of the gradient
for each input variable for the given output class.
}
\description{
This method computes the gradients of the outputs with respect to the input
variables, i.e. for all input variable \eqn{i} and output class \eqn{j}
\deqn{\frac{\partial f(x)_j}{\partial x_i}.}
}
