% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{get_activation}
\alias{get_activation}
\title{Get activation function by name}
\usage{
get_activation(name)
}
\arguments{
\item{name}{Name of the activation function.}
}
\value{
Returns the activation function.
}
\description{
This is a getter method for all implemented activation functions. Use one of
\code{"relu"}, \code{"softplus"}, \code{"sigmoid"} / \code{"logistic"},
\code{"softmax"}, \code{"tanh"}, \code{"linear"}.
}
