% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activations.R
\name{get_deveritive_activation}
\alias{get_deveritive_activation}
\title{Get deveritive of a activation function by name}
\usage{
get_deveritive_activation(name)
}
\arguments{
\item{name}{Name of the activation function.}
}
\value{
Returns the deveritive of the activation function.
}
\description{
This is a getter method for all implemented deveritives of an activation functions. Use one of
\code{"relu"}, \code{"softplus"}, \code{"sigmoid"} / \code{"logistic"},
\code{"softmax"}, \code{"tanh"}, \code{"linear"}.
}
