% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretingMethod.R
\name{get_result}
\alias{get_result}
\title{Get the result of an Interpretation Method}
\usage{
get_result(x, ...)
}
\arguments{
\item{x}{An object of the class \link{InterpretingMethod} including the
subclasses \link{Gradient}, \link{SmoothGrad}, \link{LRP}, \link{DeepLift} and
\link{ConnectionWeights}.}

\item{...}{Other arguments specified in the R6 method
\code{InterpretingMethod$get_result()}. See \link{InterpretingMethod} for details.}
}
\description{
This is a generic S3 method for the R6 method
\code{InterpretingMethod$get_result()}. See the respective method described in
\link{InterpretingMethod} for details.
}
