% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innsight_ggplot2.R
\name{+,innsight_ggplot2,ANY-method}
\alias{+,innsight_ggplot2,ANY-method}
\alias{+.innsight_ggplot2}
\title{Generic add function for \code{innsight_ggplot2}}
\usage{
\S4method{+}{innsight_ggplot2,ANY}(e1, e2)
}
\arguments{
\item{e1}{An instance of the S4 class \code{\link{innsight_ggplot2}}.}

\item{e2}{An object of class \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} or a \code{\link[ggplot2:theme]{ggplot2::theme}}.}
}
\description{
This generic add function allows to treat an instance of \code{\link{innsight_ggplot2}}
as an ordinary plot object of \code{\link{ggplot2}}. For example geoms, themes and
scales can be added as usual (see \code{\link[ggplot2:gg-add]{ggplot2::+.gg}} for more information).\cr \cr
\strong{Note:} If \code{e1} represents a multiplot (i.e., \code{e1@mulitplot = TRUE}),
\code{e2} is added to each individual plot. If only specific plots need to be
changed, the generic assignment function should be used (see
\link{innsight_ggplot2} for details).
}
\seealso{
\code{\link{innsight_ggplot2}},
\code{\link{print.innsight_ggplot2}},
\code{\link{[.innsight_ggplot2}},
\code{\link{[[.innsight_ggplot2}},
\code{\link{[<-.innsight_ggplot2}}
}
