% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innsight_ggplot2.R
\name{print,innsight_ggplot2-method}
\alias{print,innsight_ggplot2-method}
\alias{print.innsight_ggplot2}
\alias{show,innsight_ggplot2-method}
\alias{show.innsight_ggplot2}
\alias{plot,innsight_ggplot2-method}
\alias{plot.innsight_ggplot2}
\title{Generic print, plot and show for \code{innsight_ggplot2}}
\usage{
\S4method{print}{innsight_ggplot2}(x, ...)

\S4method{show}{innsight_ggplot2}(object)

\S4method{plot}{innsight_ggplot2}(x, y, ...)
}
\arguments{
\item{x}{An instance of the S4 class \link{innsight_ggplot2}.}

\item{...}{Further arguments passed to the base function \code{print} if
\code{x@multiplot} is \code{FALSE}. Otherwise, if \code{x@multiplot} is \code{TRUE}, the
arguments are passed to \link[gridExtra:arrangeGrob]{gridExtra::arrangeGrob}.}

\item{object}{An instance of the S4 class \link{innsight_ggplot2}.}

\item{y}{unused argument}
}
\value{
For multiple plots (\code{x@multiplot = TRUE}), a \link[gtable:gtable]{gtable::gtable} and
otherwise a \link[ggplot2:ggplot]{ggplot2::ggplot} object is returned invisibly.
}
\description{
The class \link{innsight_ggplot2} provides the generic visualization functions
\code{\link{print}}, \code{\link{plot}} and \code{\link{show}}, which all
behave the same in this case. They create the plot of the results
(see \link{innsight_ggplot2} for details) and return it invisibly.
}
\seealso{
\code{\link{innsight_ggplot2}},
\code{\link{+.innsight_ggplot2}},
\code{\link{[.innsight_ggplot2}},
\code{\link{[[.innsight_ggplot2}},
\code{\link{[<-.innsight_ggplot2}}, \code{\link{[[<-.innsight_ggplot2}}
}
