% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innsight_plotly.R
\name{[,innsight_plotly-method}
\alias{[,innsight_plotly-method}
\alias{[.innsight_plotly}
\alias{[[,innsight_plotly-method}
\alias{[[.innsight_plotly}
\title{Indexing plots of \code{innsight_plotly}}
\usage{
\S4method{[}{innsight_plotly}(x, i, j, ..., drop = TRUE)

\S4method{[[}{innsight_plotly}(x, i, j, ..., drop)
}
\arguments{
\item{x}{An instance of the S4 class \code{\link{innsight_plotly}}.}

\item{i}{The numeric (or missing) index for the rows.}

\item{j}{The numeric (or missing) index for the columns.}

\item{...}{other unused arguments}

\item{drop}{unused argument}
}
\value{
\itemize{
\item \verb{[.innsight_plotly}: Selects the plots from the i-th rows and j-th
columns and returns them as a new instance of \code{innsight_plotly}.
\item \verb{[[.innisght_plotly}: Selects only the single plot in the i-th row and
j-th column and returns it as a plotly object.
}
}
\description{
The S4 class \code{\link{innsight_plotly}} visualizes the results as a matrix of
plots based on \code{\link[plotly:plot_ly]{plotly::plot_ly}}. The output nodes (and also input layers)
are displayed in the columns and the selected data points in the rows. With
these basic generic indexing functions, the plots of individual rows and
columns can be accessed.
}
\seealso{
\code{\link{innsight_plotly}}, \code{\link{print.innsight_plotly}},
\code{\link{plot.innsight_plotly}}, \code{\link{show.innsight_plotly}}
}
