% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innsight_plotly.R
\name{print,innsight_plotly-method}
\alias{print,innsight_plotly-method}
\alias{print.innsight_plotly}
\alias{show,innsight_plotly-method}
\alias{show.innsight_plotly}
\alias{plot,innsight_plotly-method}
\alias{plot.innsight_plotly}
\title{Generic print, plot and show for \code{innsight_plotly}}
\usage{
\S4method{print}{innsight_plotly}(x, shareX = TRUE, ...)

\S4method{show}{innsight_plotly}(object)

\S4method{plot}{innsight_plotly}(x, y, ...)
}
\arguments{
\item{x}{An instance of the S4 class \code{\link{innsight_plotly}}.}

\item{shareX}{A logical value whether the x-axis should be shared among
the subplots.}

\item{...}{Further arguments passed to \code{\link[plotly:subplot]{plotly::subplot}}.}

\item{object}{An instance of the S4 class \code{\link{innsight_plotly}}.}

\item{y}{unused argument}
}
\description{
The class \code{\link{innsight_plotly}} provides the generic visualization functions
\code{\link{print}}, \code{\link{plot}} and \code{\link{show}}, which all
behave the same in this case. They create a plot of the results using
\code{\link[plotly:subplot]{plotly::subplot}} (see \code{\link{innsight_plotly}} for details) and return it
invisibly.
}
