% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpreting_Method.R
\name{Interpreting_Method}
\alias{Interpreting_Method}
\title{Superclass for interpreting methods}
\description{
This is a superclass for all data-based interpreting methods.
Implemented are the following methods:
\itemize{
\item Deep Learning Important Features (\link{DeepLift})
\item Layer-wise Relevance Propagation (\link{LRP})
\item Gradient-based methods:
\itemize{
\item Normal gradients (\link{Gradient})
\item Smoothed Gradients (\link{SmoothGrad})
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The given data as a torch tensor to be interpreted with the
selected method.}

\item{\code{analyzer}}{The analyzer with the stored and torch-converted model.}

\item{\code{dtype}}{The type of the data (either \code{'float'} or \code{'double'}).}

\item{\code{channels_first}}{The format of the given date, i.e. channels on
last dimension (\code{FALSE}) or after the batch dimension (\code{TRUE}). If the
data has no channels, use the default value \code{TRUE}.}

\item{\code{ignore_last_act}}{A boolean value to include the last
activation into all the calculations, or not. In some cases, the last activation
leads to a saturation problem.}

\item{\code{result}}{The methods result of the given data as a
torch tensor of size (batch_size, model_in, model_out).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Interpreting_Method$new()}}
\item \href{#method-get_result}{\code{Interpreting_Method$get_result()}}
\item \href{#method-clone}{\code{Interpreting_Method$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreting_Method$new(
  analyzer,
  data,
  channels_first = TRUE,
  dtype = "float",
  ignore_last_act = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyzer}}{The analyzer with the stored and torch-converted model.}

\item{\code{data}}{The given data as a torch tensor to be interpreted with the
selected method.}

\item{\code{channels_first}}{The format of the given data, i.e. channels on
last dimension (\code{FALSE}) or after the batch dimension (\code{TRUE}). If the
data has no channels, use the default value \code{TRUE}.}

\item{\code{dtype}}{The type of the data (either \code{'float'} or \code{'double'}).}

\item{\code{ignore_last_act}}{A boolean value to include the last
activation into all the calculations, or not. In some cases, the last activation
leads to a saturation problem.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_result"></a>}}
\if{latex}{\out{\hypertarget{method-get_result}{}}}
\subsection{Method \code{get_result()}}{
This function returns the result of this method for the given data either as an
array (\code{as_torch = FALSE}) or a torch tensor (\code{as_torch = TRUE}) of
size (batch_size, model_in, model_out).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreting_Method$get_result(as_torch = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{as_torch}}{Boolean value whether the output is an array or a torch
tensor.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of this method for the given data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreting_Method$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
