% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionWeights.R
\name{plot.ConnectionWeights}
\alias{plot.ConnectionWeights}
\title{Plot function for Connection Weight results}
\usage{
\method{plot}{ConnectionWeights}(x, rank = FALSE, scale = FALSE, ...)
}
\arguments{
\item{x}{A result of the \code{\link{Connection_Weights}} method.}

\item{rank}{If \code{TRUE}, importance scores are ranked.}

\item{scale}{Scale the importance scores to \eqn{[-1,1]}.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
Returns a ggplot2 plot object.
}
\description{
Plots the results of the \code{\link{Connection_Weights}} method.
}
\examples{
library(neuralnet)
nn <- neuralnet( Species ~ .,
                 iris, linear.output = FALSE,
                 hidden = c(10,8), act.fct = "tanh", rep = 1, threshold = 0.5)
# create an analyzer for this model
analyzer = Analyzer$new(nn)

# calculate importance scores for all classes
result <- Connection_Weights(analyzer)

# plot the results
plot(result)

# plot the results with ranked importance scores
plot(result, rank = TRUE)

# plot the results scaled to -1 to 1
plot(result, scale = TRUE)

}
\seealso{
\link{Connection_Weights}
}
