% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLift.R
\name{plot.DeepLift}
\alias{plot.DeepLift}
\title{Plot function for DeepLift results}
\usage{
\method{plot}{DeepLift}(x, rank = FALSE, scale = FALSE, ...)
}
\arguments{
\item{x}{A result of the \code{\link{DeepLift}} method.}

\item{rank}{If \code{TRUE}, contribution scores are ranked.}

\item{scale}{Scale the gradient values to the centered 90\% of the calculated
values.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
Returns a ggplot2 plot object.
}
\description{
Plots the results of the \code{\link{DeepLift}} method.
}
\examples{
library(neuralnet)
nn <- neuralnet( Species ~ .,
                 iris, linear.output = FALSE,
                 hidden = c(10,8), act.fct = "tanh", rep = 1, threshold = 0.5)
# create an analyzer for this model
analyzer = Analyzer$new(nn)

# calculate contribution scores for the whole dataset and all classes
result <- DeepLift(analyzer, data = iris[,-5])

# plot the results
plot(result)

# plot the results with ranked relevances
plot(result, rank = TRUE)

# cut the lower and upper 5\% of the data and plot the result
plot(result, scale = TRUE)

}
\seealso{
\link{DeepLift}
}
