% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRP.R
\name{plot.LRP}
\alias{plot.LRP}
\title{Plot function for LRP results}
\usage{
\method{plot}{LRP}(x, rank = FALSE, scale = FALSE, ...)
}
\arguments{
\item{x}{A result of the \code{\link{LRP}} method.}

\item{rank}{If \code{TRUE}, relevance scores are ranked.}

\item{scale}{Scale the gradient values to the centered 90\% of the calculated
values.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
Returns a ggplot2 plot object.
}
\description{
Plots the results of the \code{\link{LRP}} method.
}
\examples{
library(neuralnet)
nn <- neuralnet( Species ~ .,
                 iris, linear.output = FALSE,
                 hidden = c(10,8), act.fct = "tanh", rep = 1, threshold = 0.5)
# create an analyzer for this model
analyzer = Analyzer$new(nn)

# calculate relevance scores for the whole dataset
result <- LRP(analyzer, data = iris[,-5], rule_name = "simple")

# plot the results
plot(result)

# plot the results with ranked relevances
plot(result, rank = TRUE)

# cut the lower and upper 5\% of the data and plot the result
plot(result, scale = TRUE)

}
\seealso{
\link{LRP}
}
