% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_gradients.R
\name{wrap_torch_result}
\alias{wrap_torch_result}
\title{Wrap result in TorchGradientResult object if requested}
\usage{
wrap_torch_result(
  result,
  return_object,
  data,
  model,
  method_name,
  output_idx,
  preds = NULL,
  ...
)
}
\description{
Wrap result in TorchGradientResult object if requested
}
\keyword{internal}
