% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectionWeights.R
\name{Connection_Weights}
\alias{Connection_Weights}
\title{Connection Weights method}
\usage{
Connection_Weights(analyzer)
}
\arguments{
\item{analyzer}{An instance of the R6 class \code{\link{Analyzer}}.}
}
\value{
It returns a matrix of shape \emph{(in, out)},
which contains the importance scores for each input variable to the
output predictions.
}
\description{
This function implements the \emph{Connection Weight} method investigated by
Olden et al. (2004) which results in a feature importance score for each input
variable. The basic idea is to multiply up all path weights for each
possible connection between an input feature and the output and then
calculate the sum over them. For a neural network with \eqn{3} hidden layers with weight
matrices \eqn{W_1}, \eqn{W_2} and \eqn{W_3} this method results in a simple
matrix multiplication
\deqn{W_1 * W_2 * W_3. }
}
\examples{
library(neuralnet)
# train a neural net on the iris dataset and create analyzer
nn <- neuralnet(Species ~ ., iris, linear.output = FALSE,
                hidden = c(10,6), act.fct = "tanh", rep = 1, threshold = 0.1 )
analyzer = Analyzer$new(nn)

# calculate importance for all classes
result <- Connection_Weights(analyzer)
plot(result)

# plot the importance in a ranked scale
plot(result, rank = TRUE)

}
\references{
J. D. Olden et al. (2004) \emph{An accurate comparison of methods for
quantifying variable importance in artificial neural networks using
simulated data.} Ecological Modelling 178, p. 389–397
}
\seealso{
\code{\link{Analyzer}}, \link{plot.ConnectionWeights}
}
