% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepLift.R
\name{DeepLift}
\alias{DeepLift}
\title{Deep Learning Important FeaTures (DeepLIFT) method}
\description{
This is an implementation of the \emph{Deep Learning Important FeaTures (DeepLIFT)}
algorithm introduced by Shrikumar et al. (2017). It's a local method for
interpreting a single element \eqn{x} of the dataset concerning a reference value \eqn{x'}
and returns the contribution of each input feature from the difference of the
output (\eqn{y=f(x)}) and reference output (\eqn{y'=f(x')}) prediction.
The basic idea of this method is to decompose the difference-from-reference
prediction with respect to the input features, i.e.
\deqn{\Delta y = y - y'  = \sum_i C(x_i).}
Compared to \emph{Layer-wise Relevance Propagation} (see \code{\link{LRP}}) is the
DeepLIFT method exact decomposition and not an approximation, so we get real contributions
of the input features to the difference-from-reference prediction. There are
two ways to handle activation functions: \emph{Rescale-Rule} and \emph{Reveal-Cancel-Rule}.
}
\references{
A. Shrikumar et al. (2017) \emph{Learning important features through
propagating activation differences.}  ICML 2017, p. 4844-4866
}
\section{Super class}{
\code{\link[msIML:Interpreting_Method]{msIML::Interpreting_Method}} -> \code{DeepLift}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x_ref}}{The reference input of size (1, model_in) for the interpretation.}

\item{\code{rule_name}}{Name of the applied rule to calculate the contributions for the
non-linear part of a Neural Network layer. Either \code{"rescale"} or \code{"reveal_cancel"}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DeepLift$new()}}
\item \href{#method-clone}{\code{DeepLift$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="msIML" data-topic="Interpreting_Method" data-id="get_result">}\href{../../msIML/html/Interpreting_Method.html#method-get_result}{\code{msIML::Interpreting_Method$get_result()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of the DeepLift-Method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeepLift$new(
  analyzer,
  data,
  channels_first = TRUE,
  dtype = "float",
  ignore_last_act = TRUE,
  rule_name = "rescale",
  x_ref = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyzer}}{An instance of the R6 class \code{\link{Analyzer}}.}

\item{\code{data}}{The data for which the contribution scores are to be calculated.
It has to be an array or array-like format of size (batch_size, model_in).}

\item{\code{channels_first}}{Set the data format of the given data. Internally the format
\code{channels_first} is used, therefore the format of the given data is required.
Also use the default value \code{TRUE} if no convolutional layers are used.}

\item{\code{dtype}}{The data type for the calculations. Use either \code{'float'} or
\code{'double'}.}

\item{\code{ignore_last_act}}{Set this boolean value to include the last
activation, or not (default: \code{TRUE}). In some cases, the last activation
leads to a saturation problem.}

\item{\code{rule_name}}{Name of the applied rule to calculate the contributions. Use one
of \code{'rescale'} and \code{'reveal_cancel'}.}

\item{\code{x_ref}}{The reference input of size (1, model_in) for the interpretation.
With the default value \code{NULL} you use an input of zeros.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DeepLift$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
