% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gradien_based.R
\name{Gradient}
\alias{Gradient}
\title{Calculate the Gradients}
\description{
This method computes the gradients of the outputs with respect to the input
variables, i.e. for all input variable \eqn{i} and output class \eqn{j}
\deqn{d f(x)_j / d x_i.}
}
\section{Super classes}{
\code{\link[innsight:Interpreting_Method]{innsight::Interpreting_Method}} -> \code{\link[innsight:Gradient_Based]{innsight::Gradient_Based}} -> \code{Gradient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Gradient$new()}}
\item \href{#method-clone}{\code{Gradient$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="innsight" data-topic="Interpreting_Method" data-id="get_result">}\href{../../innsight/html/Interpreting_Method.html#method-get_result}{\code{innsight::Interpreting_Method$get_result()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gradient$new(
  analyzer,
  data,
  channels_first = TRUE,
  dtype = "float",
  ignore_last_act = TRUE,
  times_input = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{analyzer}}{The analyzer with the stored and torch-converted model.}

\item{\code{data}}{The given data as a torch tensor to be interpreted with the
this method.}

\item{\code{channels_first}}{The format of the given data, i.e. channels on
last dimension (\code{FALSE}) or after the batch dimension (\code{TRUE}). If the
data has no channels, use the default value \code{TRUE}.}

\item{\code{dtype}}{The type of the data (either \code{'float'} or \code{'double'}).}

\item{\code{ignore_last_act}}{A boolean value to include the last
activation into all the calculations, or not. In some cases, the last activation
leads to a saturation problem.}

\item{\code{times_input}}{Multiplies the gradients with the input features. This
method is called 'Gradient x Input'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gradient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
