% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradienBased.R
\name{GradientBased}
\alias{GradientBased}
\title{Superclass for Gradient-based Interpretation Methods}
\description{
Superclass for gradient-based interpretation methods. This
class inherits from \link{InterpretingMethod}. It summarizes all implemented
gradient-based methods and provides a private function to calculate the
gradients w.r.t. to the input for given data. Implemented are:
\itemize{
\item Normal Gradients and Gradient x Input (\link{Gradient})
\item SmoothGrad and SmoothGrad x Input (\link{SmoothGrad})
}
}
\section{Super class}{
\code{\link[innsight:InterpretingMethod]{innsight::InterpretingMethod}} -> \code{GradientBased}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{times_input}}{Multiplies the gradients with the input features.
This method is called 'Gradient x Input'.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GradientBased$new()}}
\item \href{#method-clone}{\code{GradientBased$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="innsight" data-topic="InterpretingMethod" data-id="get_result">}\href{../../innsight/html/InterpretingMethod.html#method-get_result}{\code{innsight::InterpretingMethod$get_result()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradientBased$new(
  converter,
  data,
  channels_first = TRUE,
  dtype = "float",
  ignore_last_act = TRUE,
  times_input = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{converter}}{The converter of class \link{Converter} with the stored and
torch-converted model.}

\item{\code{data}}{The given data in an array-like format to be interpreted
with the selected gradient-based method.}

\item{\code{channels_first}}{The format of the given date, i.e. channels on
last dimension (\code{FALSE}) or after the batch dimension (\code{TRUE}). If the
data has no channels, use the default value \code{TRUE}.}

\item{\code{dtype}}{The type of the data (either \code{'float'} or \code{'double'}).}

\item{\code{ignore_last_act}}{A boolean value to include the last
activation into all the calculations, or not. In some cases, the last
activation leads to a saturation problem.}

\item{\code{times_input}}{Multiplies the gradients with the input features. This
method is called 'Gradient x Input'.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GradientBased$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
