% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innsight.R
\docType{package}
\name{innsight-package}
\alias{innsight}
\alias{innsight-package}
\title{Make machine learning models and predictions interpretable}
\description{
Interpretability methods to analyze the behavior and
    predictions of artificial Neural Networks.
    Implemented methods are: Connection Weights described by Olden et al.
    (2004) <doi:10.1016/j.ecolmodel.2004.03.013>, Layer-wise Relevance
    Propagation (LRP) described by Bach et al. (2015)
    <doi:0.1371/journal.pone.0130140> and Deep Learning Important Features
    (DeepLIFT) described by Shrikumar et al.  (2017) <arXiv:1704.02685>.
}
\author{
\strong{Maintainer}: Niklas Koenen \email{niklas.koenen@gmail.com}

}
