% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innsight_ggplot2.R
\name{[,innsight_ggplot2-method}
\alias{[,innsight_ggplot2-method}
\alias{[.innsight_ggplot2}
\alias{[[,innsight_ggplot2-method}
\alias{[[.innsight_ggplot2}
\alias{[<-,innsight_ggplot2-method}
\alias{[<-.innsight_ggplot2}
\title{Indexing plots of \code{innsight_ggplot2}}
\usage{
\S4method{[}{innsight_ggplot2}(x, i, j, ..., drop = TRUE)

\S4method{[[}{innsight_ggplot2}(x, i, j, ...)

\S4method{[}{innsight_ggplot2}(x, i, j, ...) <- value
}
\arguments{
\item{x}{An instance of the s4 class \code{\link{innsight_ggplot2}}.}

\item{i}{The numeric (or missing) index for the rows.}

\item{j}{The numeric (or missing) index for the columns.}

\item{...}{other unused arguments}

\item{drop}{unused argument}

\item{value}{Another instance of the s4 class \code{innsight_ggplot2} but of
shape \code{i} x \code{j}.}
}
\value{
\itemize{
\item \verb{[.innsight_ggplot2}: Selects only the plots from the \code{i}-th row and
\code{j}-th column and returns them as a new instance of \code{\link{innsight_ggplot2}}.
\item \verb{[[.innsight_ggplot2}: Selects only the single plot in the \code{i}-th row and
\code{j}-th column and returns it an \link[ggplot2:ggplot]{ggplot2::ggplot} object.
\item \verb{[<-.innsight_ggplot2}: Replaces the plots from the \code{i}-th row and \code{j}-th
column with those from \code{value} and returns the modified instance of
\code{\link{innsight_ggplot2}}.
}
}
\description{
The s4 class \code{\link{innsight_ggplot2}} visualizes the results in the form of
a matrix, with the output nodes (and also the input layers) in the columns
and the selected data points in the rows. With these basic generic indexing
functions, the plots of individual rows and columns can be accessed,
modified and the overall plot can be adjusted accordingly.
}
\seealso{
\code{\link{innsight_ggplot2}}, \code{\link{print.innsight_ggplot2}},
\code{\link{+.innsight_ggplot2}}
}
