% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpretingMethod.R
\name{InterpretingMethod}
\alias{InterpretingMethod}
\title{Superclass for Interpreting Methods}
\description{
This is a superclass for all data-based interpreting methods.
Implemented are the following methods:
\itemize{
\item Deep Learning Important Features (\link{DeepLift})
\item Layer-wise Relevance Propagation (\link{LRP})
\item Gradient-based methods:
\itemize{
\item Normal gradients (\link{Gradient})
\item Smoothed Gradients (\link{SmoothGrad})
}
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The given data as a torch tensor to be interpreted with the
selected method.}

\item{\code{converter}}{The converter with the stored and torch-converted model.}

\item{\code{dtype}}{The type of the data (either \code{'float'} or \code{'double'}).}

\item{\code{channels_first}}{The format of the given date, i.e. channels on
last dimension (\code{FALSE}) or after the batch dimension (\code{TRUE}). If the
data has no channels, use the default value \code{TRUE}.}

\item{\code{ignore_last_act}}{A boolean value to include the last
activation into all the calculations, or not. In some cases, the last
activation leads to a saturation problem.}

\item{\code{result}}{The methods result of the given data as a
torch tensor of size (batch_size, dim_in, dim_out).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{InterpretingMethod$new()}}
\item \href{#method-get_result}{\code{InterpretingMethod$get_result()}}
\item \href{#method-plot}{\code{InterpretingMethod$plot()}}
\item \href{#method-clone}{\code{InterpretingMethod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of this class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InterpretingMethod$new(
  converter,
  data,
  channels_first = TRUE,
  dtype = "float",
  ignore_last_act = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{converter}}{The converter with the stored and torch-converted model.}

\item{\code{data}}{The given data in an array-like format to be interpreted
with the selected method.}

\item{\code{channels_first}}{The format of the given data, i.e. channels on
last dimension (\code{FALSE}) or after the batch dimension (\code{TRUE}). If the
data has no channels, use the default value \code{TRUE}.}

\item{\code{dtype}}{The type of the data (either \code{'float'} or \code{'double'}).}

\item{\code{ignore_last_act}}{A boolean value to include the last
activation into all the calculations, or not. In some cases, the last
activation leads to a saturation problem.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_result"></a>}}
\if{latex}{\out{\hypertarget{method-get_result}{}}}
\subsection{Method \code{get_result()}}{
This function returns the result of this method for the given data
either as an array (\code{'array'}), a torch tensor (\code{'torch.tensor'}) of
size (batch_size, dim_in, dim_out) or a data.frame (\code{'data.frame'}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InterpretingMethod$get_result(type = "array")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The data format of the result. Use one of \code{'array'},
\code{'torch.tensor'} or \code{'data.frame'} (default: \code{'array'}).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The result of this method for the given data in the chosen
format.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
This method visualizes the result of the selected method in a
\link[ggplot2:ggplot]{ggplot2::ggplot}. You can use the argument \code{data_id} to select
the data points in the given data for the plot. In addition, the
individual classes for the plot can be selected with the argument
\code{class_id}. The different results for the selected data points and
classes are visualized using the method \link[ggplot2:facet_grid]{ggplot2::facet_grid}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InterpretingMethod$plot(data_id = 1, class_id = 1, aggr_channels = sum)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_id}}{An integer vector containing the numbers of the data
points whose result is to be plotted, e.g. \code{c(1,3)} for the first
and third data point in the given data. Default: \code{c(1)}.}

\item{\code{class_id}}{An integer vector containing the numbers of the classes
whose result is to be plotted, e.g. \code{c(1,4)} for the first and fourth
class. Default: \code{c(1)}.}

\item{\code{aggr_channels}}{Pass a function to aggregate the channels. The
default function is \link[base:sum]{base::sum}, but you can pass an arbitrary function.
For example, the maximum \code{max} or minimum \code{min} over the channels or
only individual channels with \code{function(x) x[1]}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a \link[ggplot2:ggplot]{ggplot2::ggplot} with the plotted results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InterpretingMethod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
